[KYUUBI #1872] Make ProcBuilder decouple with YARN
